Lee/demo behavior chooser

+++ basestation/include/anki/cozmo/basestation/behaviors/behaviorOCD.h
@@ -36,14 +36,14 @@ namespace Cozmo {
     BehaviorOCD(Robot& robot, const Json::Value& config);
     virtual ~BehaviorOCD() { }
     
-    virtual bool IsRunnable(float currentTime_sec) const override;
+    virtual bool IsRunnable(double currentTime_sec) const override;
 
     virtual Result Init() override;
     
-    virtual Status Update(float currentTime_sec) override;
+    virtual Status Update(double currentTime_sec) override;
     
     // Finish placing current object if there is one, otherwise good to go
-    virtual Result Interrupt(float currentTime_sec) override;
+    virtual Result Interrupt(double currentTime_sec) override;
     
     virtual bool GetRewardBid(Reward& reward) override;
